Adicionando menu professor e funcionário

+++ back-end/src/main/java/spge/spge/dto/request/DefinirNotaRequestDTO.java
@@ -11,5 +11,5 @@ public class DefinirNotaRequestDTO {
     private Long codigoAluno;
     private Integer numeroDoBimestre;
     private String materia;
-    private Integer nota;
+    private Double nota;
 }
